"4:30pm. https://www.reddit.com/r/ProgrammingLanguages/comments/adtp7l/d_what_are_good_choices_for_interpreter/edkc1nu/?st=jqnxmgme&sh=fda99afd
"I've found hashmaps to be too slow. Maybe that was because of the hashing algorithm I was using (robin hood hashing with Rust's hashmap), but if you're accessing the same variable in a long loop, you're going to have some significant overhead of having to hash the same string over and over.

Now I opt for slot-based variable access like the OP mentioned - number all of your variables, and shove them in a vector."

I think Java caches the string hash, so I assumed that .NET does too, but I've never really checked.

5:15pm. "> you're going to have some significant overhead of having to hash the same string over and over.

I've heard that Java caches the string's hash in the object itself so I assumed that .NET does so as well, but I've never really [checked](https://stackoverflow.com/questions/3053726/why-might-a-system-string-object-not-cache-its-hash-code). Going through that SO answer is definitely a strong point in favor of numbering the variables.

Edit: Choice quote.

> I guess my point is that for the author(s) of this implementation to worry, "What if this string instance is modified between calls, even though the class as it is publicly exposed is immutable?" would be like for someone planning a casual outdoor BBQ to think to him-/herself, "What if someone brings an atomic bomb to the party?" Look, if someone brings an atom bomb, party's over."

My reply is a bit of a non-sequitor, but whatever.

https://andrewlock.net/why-is-string-gethashcode-different-each-time-i-run-my-program-in-net-core/

Right now I am reading this.

Well, I really won't be doing any programming just yet. Let me review the source for a bit and then I will call it a day.

Today I did nothing except worry and slack. Tomorrow I'll want to do some focused work.

The real work I did today was during the night. Right now I am still in planning mode.

One change I am going to have to introduce is to have separate type for modules and functions. Right now they are mostly the same thing, but once I move to `PersistentVector`, I'll have to use different means to deal with it.

https://fsprojects.github.io/FSharpx.Collections/reference/fsharpx-collections-persistentvector-1.html

Ah, the Conj for it really is O(1). That is really great.

https://github.com/fsprojects/FSharpx.Collections/blob/master/src/FSharpx.Collections/PersistentVector.fsi

Oh, it has fold, map and ...not filter. But that I can easily implement using `init` for the particular use case that I need.

Besides this, I am going to use two environments, one for the real thing and the other for those temporary pattern variables. I do not want temporaries to get stuffed into the main environment as now I won't be doing filtering on every binding. Doing filtering on every binding is just so insane from a performance standpoint that I am amazed at myself for going through with it.

6:30pm. Done with lunch. Did some more replying in that thread.

https://github.com/fsprojects/FSharpx.Collections/blob/master/src/FSharpx.Collections/PersistentVector.fs#L254

I was wondering where the immutable updates were - it seems `TransientVector` is the mutable counterpart to `PersistentVector`.

6:40pm. Let me stop here for the day. I think I've planned it out enough. Tomorrow I will roll up my sleeves and get dirty.

None of the changes I will introduce will really be difficult to implement. Changing how pattern matching works will break all the tests so I'll leave that for later."