"10am. I got up an hour ago and did some Reddit posting. The latest post might turn out to be inflammatory.
https://karczmarczuk.users.greyc.fr/TEACH/Doc/EssProgLan.pdf
Essentials Of Programming Languages

Let me take a look at the Essentials of programming languages which was recommended to me yesterday. I'll save entertaiment for later.

Before do anything else I want to make a list of things I need to do for the language. I keep having ideas and putting them in one place will make it easier to organize me.

---

* Restrict `open` to global modules only.
* Redo the prepass so it does variable numbering. Doing this properly will require phased compilation as `open`s will have to be evaluated during the prepass.
* Replace variable filtering with liveness tests in renaming and equality tests.
* Replace the immutable map for function environments with persistent vectors.
* Remove common subexpression elimination from the partial evaluator.
* Separate uncase from pattern matching.
* Change pattern matching from CPS'd form to linear.
* Do destructuring directly after join points rather than during bindings. Introduce a new `TypedExpr` to streamline this.

---

The first 4 are the core language related changes that I have planned.

Removing CSE and separating uncase are there to remove the patterm matching related confusion where annoyingly the term outside the scope will get rewritten to the uncased expression.

Also, right now there are various places where I do `match x with () | _ -> ...` just to uncase a boxed type. This is really annoying and is telling me I went wrong somewhere.

This will also allow me to change pattern matching from CPS'd to linear. I want this so I can take care of that annoying bug scoping bug that is described in the manual.

10:25am. The last thing is I can stop worrying on when to `destructure`. I already understand that I should be doing this at join points, but the reason why I am not doing this now is because it would block tail call optimization. It is really annoying and I should have done it like I will now a long time ago.

10:35am. I am still thinking about it.

Yeah, I guess I cannot go for such huge changes right of the bad. I need to digest them a bit first in mind.

I do not think it will be too bad that once I am done with this I will drop Spiral for a while. I've indulged myself a lot over the past 4 years which the trajectory I've pursued, but it will all work out fine in the end.

10:40am. It is fine to move to climbing a different mountain once one is done. It is fine to achieve a partial victory now in order to ensure a complete victory at a later point.

I am actually looking forward to doing Python programming for a change now as I want to experience the power of dynamic languages, but I want the basic Spiral interpreter to be perfect before I leave to do that. There is just too much unfinished business left in Spiral.

Let me read the book I linked to for a while first."