"10:50am. "The notion that there was some underlying structure to the way the language was organized, and that I might want to override some of the language designers’ decisions, never occurred to me."
This used to be my mindset as well.

11:05am. https://boards.4channel.org/a/thread/183082527/love-live-sunshine

Focus me. Let me just skim this book.

11:20am. No, Spiral would not do well as a dynamic language in the naive case where it boxed everything. All the partial application would make it incredibly inefficient.

A proper implementation using bytecode would do much better.

...But wait. How would I box inlinleables then?

11:30am. Actually would this be as inefficient as I think in a dynamic language?

Well, the issue would be how much destructuring and structuring it does.

Using boxed types in STLC mode would not work, but in dynamic mode I pass environments around directly and relax my restrictions on flat representations a little.

In dynamic mode there is one issue with that vision though. Although it would work there would be no point join points then, and there would be no point in inlining.

11:35am. How would inlineables work for dynamic languages?

11:55am. Well, they would not. If you wanted to do metaprogramming in a dynamic language in order to achieve stage separation it might be better to in fact have Lisp style macros.

Also I realized how I could do complete type inference for Spiral for data structures. I could have them act as achors. An assignment to them that would be of the wrong type would raise an exception to that point and the program would be reevaluated with the new type.

This similar to `type_raise`, `type_catch` that I did one point in the language. Although to be honest, it might make more sense to just use a different language rather than do that. Spiral's compile times are crazy enough as they are already.

12pm. Let me get breakfast.

1:10pm. I get it now. I think Lisp style macros give advantages that are somewhat intrinsic to dynamic languages. Actually, they have some advantages everywhere, but the place where they are definitely the most pronounced are in Lisp-style languages.

"To turn the argument from this post around, if functional macros of the kind Spiral favors are so good then why don't Lisps use them? I am trying to think of ways doing this right now and it occurs to me that they would be useless in a dynamic language. It would be quite difficult to get any stage separation and after the implementation is done they'd just end up being regular functions.

On the other hand, how to get the same metaprogramming facility with Lisp-style macros in dynamic languages is relatively transparent.

So maybe there is a very real point to Lispy macros that I've overlooked."

I am a bit uncomfortable about the post that I made because it not like I intended to go out of my way to troll here. I am just thinking things through.

In practice though when it comes to performance it is not about macros. One pattern that dynamic languages like Python have converged in order to get good performance is calling libraries written in high performance ones.

1:25pm. I just can't relax right now. I've been reading the book on the side (it is very dry) and my thoughts are all about Spiral. I am done with the book, I had enough of it.

I actually hadn't take a proper break during breakfast.

As a ideal language design I am really converging to the `dynamic -> intensional -> simple`. Spiral's `intensional` is just too hard, but on its own dynamic languages are too slow. And they could do much better than calling C libraries.

Higher level systems languages than C on the other hand are not good at being library languages.

I know this for a fact about C++. This is why C and Fortran persist to this day.

1:40pm. https://www.youtube.com/watch?v=-mhBD8Frkc4&feature=youtu.be&t=629

Let me watch this for a bit. I feel a really deep unrest right now. I can't program.

I think I'll get away from the screen for a bit in order to digest my desires properly.

1:45pm. I had enough. Let me do the chores and I'll sleep in for a while."