"4:15pm. Holy shit. 2.5 hours?
I've come to a decision - Spiral needs to become an embeddable library language similar to some interpreted dynamic languages.

Spiral's very small size makes is highly suitable for such a role.

I want to be able to call Spiral from Python. I want it to kill PyTorch Script. I want to be able to make a game in Spiral that calls Python out for deep learning services than in turn calls Spiral to compile the functions for its backend. On top of that I want Python to handle the game's process.

Programs communicating with other programs is not exactly a stretch of imagination. I want Spiral to be a caching interpreter. I want it to be an service that other languages can tap into.

And I do not want to redo it in Python or in anything else. F# itself should suffice for such a role.

After it is done, I should be able to compile it to native code without dependencies on .NET.

4:25pm. Getting this sort of thing done efficiently should be a fair goal.

This was actually Spiral's original goal. Back in 2016 I really did not want to make Spiral for obvious reasons - it would take too long. Eventually I fell into it, but at first I wanted to make it so that F# is dynamically capable of calling the Spiral compiler.

Along the way that fell to the wayside, but now it is time to make it a priority.

It is not like my reasons for making Spiral are dead just because it won't be good for frontends, GUIs, interactive programming and other good things.

I am still just too green and immature. Eventually all algorithms find their purpose. Dynamic interpreted languages are never going to just die out.

Just like in 2016 with F#, right now I am really getting slammed by just how much important expressiveness is. It is important in ways I could not even imagine. I never thought that operating from an interpreted regime could give so many benefits.

I made Spiral to effectively communicate with the GPU because crossing language boundaries was so difficult in F#. Right now I am dealing with this again except on a whole another level.

Even if I went to Python, of course that on its own it will not be enough. Eventually it will be necessary to drop down to a lower level. And when the need to that arises, Spiral will be there to keep me safe.

I cannot keep myself from falling into the abyss, but before I jump I'd better bring along a parachute.

4:40pm. I feel like learning more about interpretation, but before I'll do that as I go along. I do not feel like studying using dry books. Learning is best done piece by piece. I'll keep `Essentials of Programming Languages` in mind as a reference, though I did not get much use from books in the last four years.

I do not think I will be doing any programming today.

Tomorrow absolutely, I will start work on the new Spiral.

Since it will have role to play, I should not hesitate in putting more effort into it. Without a doubt, I will find a way to draw out the full power of both dynamic and intensional languages. I will not leave any of the aspects behind in my search. Spiral is not just that weak that it would be left behind by any other language.

My victory if it comes will come with dynamism and intensionality working - together."