"11am. I've actually forgotten quite a bit about the language internals by this point. Having a type system is really helpful here. If this was in a dynamic language I'd be in trouble now.
```
[<CustomComparison;CustomEquality;StructuredFormatDisplay("{AsString}")>]
type ConsedNode<'a> =
    {
    node: 'a
    tag: int
    hkey: int
    }
```

Why is there difference between `tag` and `hkey` in HashConsed?

11:15am.

```
and MapType =
    | MapTypeFunction of FunctionCore // Type level function. Can also be thought of as a procedural macro.
    | MapTypeRecFunction of FunctionCore * string
    | MapTypeModule
```

I've done a basic review of the types. Now it is time to break everything. I'll leave module as is for now, but functions will use the PersistentVector from here on out.

11:20am.

```
and PersistentVector<'a> = {
    all : FSharpx.Collections.PersistentVector<'a>
    live : int[]
    }
and EnvTy = PersistentVector<Ty>
```

I need to make a specific class for this that implements custom hashing and equality.

...Actually, how about I inherit from it instead of composing with it? That should be more efficient.

11:35am.

```
type EnvVector<'a when 'a : equality and 'a : comparison>(all : FSharpx.Collections.PersistentVector<'a>, live_indices : int[]) = // `live` is assumed to be ordered and immutable
    member x.Live = live_indices |> Array.map (fun x -> all.[x])
    override x.ToString() = "<vector>"
    override x.GetHashCode() = hash x.Live
    override x.Equals(y) = // TODO: Make this more efficient.
        match y with
        | :? EnvVector<'a> as y -> x.Live = y.Live
        | _ -> failwith "Invalid equality for EnvVector."

    interface IComparable with
        member x.CompareTo(y) =
            match y with
            | :? EnvVector<'a> as y -> compare x.Live y.Live
            | _ -> failwith "Invalid comparison for Node."
```

This should suffice for now. First I'll make it all work and then I'll optimize it. It really has been a while since I did this last. It really would suck to have to run the thing in order to do type checking first.

11:40am.

```
// For Common Subexpression Elimination. I need it not for its own sake, but to enable other PE based optimizations.
and CSEDict = Map<TypedExpr,TypedExpr> ref
```

I'll remove this thing.

11:45am.

```
| TyJoinPoint of JoinPointKey * JoinPointType * Arguments * Ty
```

Ah, I already have this. Amazing that it did not occur to me take advantage of this in destructure. Actually, maybe I should rename this to TyJoinPointCall and add TyJoinPointDestruct?

No. Let me give it some special handling.

```
| TyJoinPoint of JoinPointKey * JoinPointType * Arguments * TypedExpr * Ty
```

That `TypedExpr` will be used to make a list of bindings to which the thing was destructed.

```
| TyJoinPoint of JoinPointKey * JoinPointType * Arguments * DestructedArgs * Ty
```

Let me alias it like this.

12pm.

```
let c = function
| Env env -> env
| EnvConsed env -> env.node

let (|C|) x = c x
let (|CN|) (x: ConsedNode<_>) = x.node
```

Now it is this thing that is being annoying.

In order to make things fast, how about I forget all about join points and layout types?

Just forget about memoizing the expressions and all that complicated stuff.

For v0.1 how about I just focus on making it as fast as possible in normal mode.

12:05pm. I think I will make a separate `TypedExpr` just for `Consed` data types.

```
and ConsedEnvTerm = EnvVector<ConsedTypedData>
and ConsedMapTerm = Map<string,ConsedTypedData>

and ConsedTypedData =
    | CTyT of Ty
    | CTyV of TyTag
    | CTyBox of ConsedTypedData * Ty
    | CTyList of ConsedNode<ConsedTypedData list>
    | CTyVector of ConsedNode<ConsedEnvTerm>
    | CTyMap of ConsedNode<ConsedMapTerm>
    | CTyLit of Value
```

This will be for the join points and layout types.

This will allow me respite from having to constantly check whether the current environment is consed or not.

Now...I just remembered, but are Exprs always nodified?

```
and [<CustomEquality;CustomComparison>] Expr =
    | V of string * tag: int
```

I flattened it. Ok, fine, I'll leave this as is.

12:15pm.

```
and TraceNode<'a when 'a:equality and 'a:comparison>(expr:'a, trace:Trace) =
    member x.Expression = expr
    member x.Trace = trace
    override x.ToString() = sprintf "%A" expr
    override x.GetHashCode() = hash expr
    override x.Equals(y) =
        match y with
        | :? TraceNode<'a> as y -> expr = y.Expression
        | _ -> failwith "Invalid equality for TraceNode."

    interface IComparable with
        member x.CompareTo(y) =
            match y with
            | :? TraceNode<'a> as y -> compare expr y.Expression
            | _ -> failwith "Invalid comparison for TraceNode."
```

Am I using this anywhere at all?

No, it does not seem so. Removed.

12:20pm.

```
and get_type = function
    | TyLit x -> get_type_of_value x
    | TyList l -> List.map get_type l |> ListT
    | TyMap(C l, t) -> MapT (env_to_ty l, t)
```

I'll take care of this later. Let me stop here so I can eat. I am starting to get into it now."