"10:45am. https://www.reddit.com/r/reinforcementlearning/comments/aese8o/ilya_sutskever_openai_i_see_no_worthy_games_for/
"I sort of agree, but with some caveats. You have to keep in mind that OpenAI tends to be fairly extreme in its views.

1) The most important caveat is that there is a big difference between say solving a game and solving it properly. Deepmind is the biggest culprit here - they 'solved' Go, but not in a particularly computationally effective manner.

The original Deep Blue took a supercomputer, and now it would be beat by a mobile phone AI mostly because of better heuristics underneath. That is still yet to happen for Go. The strategy for baking in such prior knowledge into deep RL is not obvious as it is for minimax algorithms.

2) Adversity and being able to self play makes everything so much easier as it gives a convenient target to optimize. Not all games offer a convenient adversary to fight against. I am thinking of single-player games in particular, but not AAA ones specifically.

There is a huge gap in our understanding how to tackle this sort of thing. It all goes back to how nobody knows at this point how to do unsupervised learning effectively. 'Unsupervised' or 'metalearning' or 'hierarchical' are just stopgap words for things we do not understand.

Maybe Ilya thinks that solving Dota 2 will require solving unsupervised learning, in which case, yes, then just solving Dota 2 will be enough.

Their previous attempt at Dota 2 though was to use PPO + moon sized supercomputer. Actions speak louder than words and I hope their next showing will be better. I hope to see better algorithms instead of massive compute. Right now they are acting as if the naive interpretation of their views is the right one.

Though it would not get them as much media attention and hence investor money, if OpenAI is really serious about their task they should backtrack a little and take on the ambitious goal of improving upon the sample complexity of AlphaZero by a few orders of magnitude. They are going to need to be able to do that anyway if they really intend to beat Dota.

Or they could just get a Jupiter sized computer next time and see how that works out. It is their call."

I just felt like posting this.

https://arxiv.org/abs/1812.07297
Continual Match Based Training in Pommerman - Technical Report

I'll save this for later. I want to get some programming done. I should not have slacked so much yesterday.

I can't believe how late I got up today because of that. But it is fine. I was hard at thought on what to do next though.

...I forgot to push the changes I made yesterday.

11:05am. Let me have breakfast here. Rather than start now, it makes more sense to get it out of the way first. Then I'll go at it full force.

Why am I slacking so much when I am actually looking forward to programming? It is a mystery. I get into it once I start, but in the interim the less said the better."