"1:10pm. Done with breakfast. Enough of this, I need to move forward.
1:15pm. While I was having breakfast I took a look at `PersistentVector`'s internals.

Then I got the sense that I won't be able to meet my performance goals.

That `ref None` inside. That let statement. That `obj []`. In order to get really good interpreted performance, I need to get the bytecode interpreter as close as possible to the machine code. That means doing the low level stuff in C, or at least making Spiral in Spiral.

This is absolute - for Spiral to reach its true potential and usefulness, it has to be done perfectly across the board.

1:30pm. It is difficult to describe the sheer and absolute terror I feel at that thought. It is as if my life is literally in danger.

```
* Restrict `open` to global modules only.
* Redo the prepass so it does variable numbering. Doing this properly will require phased compilation as `open`s will have to be evaluated during the prepass.
* Replace variable filtering with liveness tests in renaming and equality tests.
* Replace the immutable map for function environments with persistent vectors.
* Add an extra environment for temporaries.
* Separate uncase from pattern matching.
* Do destructuring directly after join points rather than during bindings. Introduce a new `TypedExpr` to streamline this.

* Check `type_catch` for correctness.

(Optional)
* Replace pattern matching on the AST with indirect threading.
* Introduce let statements.
```

Even if I ignore not being able to meet my speed targets, there is just so, so much work to be done here.

Yeah, turning Spiral into an embedded language is certainly the right thought. It should be `dynamic -> intensional -> simple`. And like Lisp demonstrates, there should be higher level of meta programing on top. I am partial to partial evaluation and would do it using function, but either will do.

Building large towers of abstraction and collapsing them down to machine code is the right way to do this.

Spiral definitely has potential and needs to be perfected, but the workload ahead of me is too much. I do not have any more time. Not even as much as a single month.

I planned out the next version, but no longer have wherewithal to play with it like I did for the last two years.

It would be different if I could guarantee a 10-100x increase in compilation time, but that won't be the case. I am afraid that I am heading for another failure. It would really clean up the internals, but that is not enough to justify the time spent working on it.

1:40pm. I am going to stop development on it here for the time being and will pick it up once I am done with the first ML arc.

I need to go for Python after all. Eventually, the need to build towers will come and Spiral will find its role as the last step in the whole system.

Ahhh, truly it would be so great if my ability was such that I could do in a day what would otherwise take me a month now. I regret it, I regret it so badly.

Last night I had dreams about doing code transformations and the benefits it would bring. It was so beautiful. If only that sort of programming could be made much more productive, it would be such a strong power.

That is without a doubt the Spiral Dream.

1:55pm. ...Goodbye Spiral. The battle that have been fought will not have been in vain.

When the time comes to resume, there will be reinforcements along for the ride."