"8:25pm. Ok, I see it. Without CSE in the partial evaluator it would be possible to get efficient pattern matching by doing partial evaluation in the prepass. It would literally be by doing CSE. I could do it. I can see how to do it now.
And this would allow me to do nested pattern matching without removing the uncase from patterns.

This could really be amazing. I can already see the ending to this.

8:35pm. There is something I do not get.

I think my issues with current CSE stem from the fact that it is `TypedExpr -> TypedExpr`. If I made it `TypedOp -> TyTag` then it would be an issue either way.

8:40pm.

```
and TypedOp =
    | TyOp of Op * TypedData []
    | TyIf of tr: TypedBind [] * fl: TypedBind []
    | TyCase of (TypedData * TypedBind []) []
    | TyJoinPoint of JoinPointKey * JoinPointType * CallArguments
```

Yeah, it could work. It would also make it easy tell which things to ignore and which to allow. I should allow `TyOp` in CSE rewriting, but keep the rest.

8:50pm. Actually, I do not even have to do the partial evaluation during the prepass. Changing the CSE functioning to something more sensible will work the magic all on its own.

```
| TyOp of Op * TypedData
```

I think I will remove the array requirement from this.

Actually, I could make it...

```
and TypedOp =
    | TyOpUnary of Op * TyTag
    | TyOpBinary of Op * TyTag * TyTag
    | TyIf of tr: TypedBind [] * fl: TypedBind []
    | TyCase of (TypedData * TypedBind []) []
    | TyJoinPoint of JoinPointKey * JoinPointType * CallArguments
```

No actually, let me leave it as...

```
and TypedBind =
    | TyLet of TypedData * TypedOp * Ty * Trace
    | TyState of TypedOp * Ty * Trace

and TypedOp =
    | TyOp of Op * TypedData
    | TyIf of tr: TypedBind [] * fl: TypedBind []
    | TyCase of (TypedData * TypedBind []) []
    | TyJoinPoint of JoinPointKey * JoinPointType * CallArguments
```

Doing this will let me remove the `TyReturn`. That thing should be an `Op` anyway.

And doing this will let me set the type of the CSE rewrite function to `Op * TypedData -> TyTag`...actually, no it has to be `Op * TypedData -> TypedData`. It needs to be whatever the uncase and hence the environment can bind to.

9:15pm. There is something I could add to the language. The `IfFail`. It tests for true and if false calls the fail case with the type of the positive.

...Forget that for now.

```
| TyCase of TyTag * (TypedData * TypedBind []) []
```

This should also have a `TyTag`.

9:20pm.

```
and TypedBind =
    | TyLet of TypedData * TypedOp * Ty * Trace
    | TyState of TypedOp * Ty * Trace

and TypedOp =
    | TyOp of Op * TypedData
    | TyIf of tr: TypedBind [] * fl: TypedBind []
    | TyCase of (TypedData * TypedBind []) []
    | TyJoinPoint of JoinPointKey * JoinPointType * CallArguments
```

Let me just ask this at this point. Do I really need `TyState`? I think it would be fine if I let TyLet bind to an empty List to indicate that it is a statement or a return.

9:30pm.

```
and TypedOp = // TypedData being `TyList []` indicates either a statement, or a local return if it is in last place of the array.
    | TyOp of TypedData * Op * TypedData * Ty * Trace
    | TyIf of TypedData * tr: TypedOp [] * fl: TypedOp [] * Ty * Trace
    | TyCase of TypedData * TyTag * (TypedData * TypedOp []) [] * Ty * Trace
    | TyJoinPoint of TypedData * JoinPointKey * JoinPointType * CallArguments * Ty * Trace
```

It is not pretty, but it will get the job done.

Let me pretty it up a bit more.

9:35pm.

```
and TypedExpr = // TypedData being `TyList []` indicates either a statement, or a local return if it is in last place of the array.
    | TyOp of TypedData * Ty * Trace * Op * TypedData
    | TyIf of TypedData * Ty * Trace * tr: TypedOp [] * fl: TypedOp []
    | TyCase of TypedData * Ty * Trace * TyTag * (TypedData * TypedOp []) []
    | TyJoinPoint of TypedData * Ty * Trace * JoinPointKey * JoinPointType * CallArguments
```

Yeah, it is better to cleanly separate the bind and the op on two sides rather than sandwiching the op.

9:40pm.

```
    | TyFunction of EnvTerm * FunctionCore
    | TyRecFunction of EnvTerm * FunctionRecCore
```

Just how is the comparison for this done. Not by comparing the environments first I hope.

```
    | TyFunction of FunctionCore * EnvTerm // TODO: Make sure that environments are being compared last.
    | TyRecFunction of FunctionRecCore * EnvTerm
```

I'll ask this question on the SO tomorrow. Right now I am too tired.

9:45pm. Let me finally call it a day.

I think that one the next redesign I have a good chance of really nailing down what Spiral truly is.

I said goodbye, but I guess maybe I will find the time to work on it on the side."