"9:35am. I am up. As a result of all my thinking tonight, I've managed to figure out what representation I want for closures. It would have to be flattened ones that just use arrays. Lists of arrays might be good for regular interpreters, but in Spiral they'd be way too much a nuisance due to the need to do renaming and equality testing everywhere. I'd have to keep converting to the flattened representation.
I do not think that flattened representations would be too much of a problem. For one, I am currently doing this except much more inefficiently. So by simple reasoning, I can in fact infer that doing this conversion would significantly improve the performance of the compiler.

---

(Type)
* Make it so CSE rewrite function is `Op * TypedData -> TypedData`.
* Replace the immutable map for function environments with arrays.
* Replace nodification on `Expr`s with reference equality. Remove the useless tag fields. It seemed a good idea at the time, but every `Expr` needs to be unique due to holding positional information.
* Do destructuring directly after join points rather than during bindings. Make `TyLet` more flexible in order to enable this.

(Prepass)
* Restrict `open` to variables in the previous compilation phases only.
* Add a let conversion prepass.
* Add a variable numbering prepass. Doing this properly will require phased compilation as `open`s will have to be evaluated during the prepass. This one would be essential for the new array representation.
* Add a register allocation prepass. Flattened array representation would help here as unlike with a list of arrays the analysis will be easy.

(Auxiliary)
* Replace pattern matching on the AST with indirect threading in the partial evaluator.
* Check `type_catch` for correctness.

---

So my plan has changed significantly since just a few days ago. It is actually a good thing I did not start, as I am definitely not done planning the next Spiral version. I guess I'll do it when the model of it settles down.

10am. I guess I'll get back to Python after chilling a bit. I am already faltering in my determination to leave Spiral alone. So far my impression of Python is negative again - not so much because of it as a language, but because all the PyTorch and Gym examples are broken in some way. What the hell are those guys doing?

The data_loading example just ends with a broken pipe exception and displaying the graphics leads to the window freezing. Yesterday I spent the whole day just trying to figure out how to make the debugger work.

Today my plan is to play around with GUIs in Python just get a sense for that freezing issue in the plotting library."