"9:30am. https://smalltalkexpress.quora.com/Why-Perl-Succeeded-and-Smalltalk-Failed https://www.quora.com/What-are-the-differences-between-Smalltalk-and-Ruby https://smalltalkexpress.quora.com/Why-Perl-Succeeded-and-Smalltalk-Failed https://blog.arkency.com/2017/01/why-smalltalk-instead-of-ruby/
9:45am. Yeah, I could study. Or I could read this stuff.

9:50am. Let me finally start reading the book.

11:35am. 50/376. Let me stop here. I got through the first chapter.

12:45pm. Let me resume. Right now I am just wasting time reading social media.

12:55pm. I'll figure out how to post commits using Pharo later. I've opened a new repo on Github.

Let me go through the book now.

1:05pm. Oh, it seems the book will literally be having me create cells to click on. Score. This is exactly what I wanted. This is making it a lot more likely that my choice is the right one.

1:15pm. I need to focus. I am too distracted right now. Let me put in a solid few hours into this thing. I'll get through the book in the next few days.

2:10pm. "The arguments go right after the colons. If you are used to languages that put all of the arguments together inside parentheses, this may seem weird at first. Don’t panic, it’s only syntax! It turns out to be a very good syntax because the name of the method can be used to explain the roles of the arguments. For example, it is pretty clear that Matrix rows: 5 columns: 2 has 5 rows and 2 columns, and not 2 rows and 5 columns."

I really do like this sort of syntax.

3:15pm. The Java influence is clear here, except with Pharo they come out and say that the IDE is needed to program in it outright. It is really a pity Java is the one that got popular instead of this.

Let me take a short pause here. I need to do the chores. I am nearly done with the chapter, I'll finish it when I get back.

3:55pm. Phew, I am done. Let me resume.

4:35pm. I am trying to register a toy project on `smalltalkhub.com` and it is telling me wrong username or password. This is nice, except there are no username or password fields! And the thing literally ate my fields including the description I wrote out before asking again. Is there something wrong with the site right now?

I think I'll just find a way to post on Github directly.

5:10pm. Let me resume now that the lunch is over. Let me finally finish the chapter.

...No actually let fire a mail to Esug.

5:20pm. Done. Let me finish the chapter. I'll come back to this part of it when the issue gets resolved.

It seems there is a blog post on how to use Pharo with Git. I'll take a look at it.

6:25pm. Let me stop here. I had enough for one day.

Though the hands on chapter was rather simple, there was actually a lot to internalize and it exhausted me.

For the past week or so, my posting has been interspersed between Python Projects and Spiral. I'll post this one on the Spiral language repo. I probably won't post on the Python Projects one.

I gave Python a decent chance, but it botched it with the GUIs and plotting. There is no way I am going to be doing ML work from the command line. Live coding will be a necessity for what is to come.

I am betting that it will work that much better when the objects are literally alive."