"9:25am. It was interesting how the Java update actually offered to uninstall itself since I haven't used it for over 6 months. That is classy.
9:30am. Well, let me resume studying Pharo. I've gotten a little taste of it yesterday. The programming that I did is not in any way difficult, but it costs a lot of energy to internalize a whole new programming environment.

With regards to Pharo itself, I feel that it has some archaic things like explicit returns and variable declarations, but if that sort of thing is needed for the IDE then I will accept it. As that talk said, the language as simple as it would not be worth it without an IDE, but with an IDE it is in a league of its own.

```
true ifTrue: [2] ifFalse: [3]
```

Its syntax is a miracle of language design - nearly the opposite of Lisp. Expression based ifs are really one of the things that indicates its authors actually understand programming unlike Pythoners. Whenever I see assignment where functional programming will do, I think how stupid and wasteful that sort of behavior is.

9:40am. But really, going into exile once again at this point...just how long will doing all of this properly take me? I wish I had started this a decade earlier. Then it would be all done by now.

9:45am. Let me finish the Tonari thread and then I will resume the book.

Also I do not feel like resuming programming on Spiral as I am into Pharo now. I'll do it next month or something once I get more intimate with Pharo."